Add a Go solution for Problem3
Use a basic sieve of Eratosthenes to generate a list of primes that could be factors of the limit, and check whether each one is a factor in the main thread.

Hugely computationally expensive, we're doing a ton of unnecessary work and I'm not a fan of implementing a prime sieve as a daisy-chained set of channels/goroutines, it uses too much memory IMO.

I have ideas on how to rewrite this to make it faster but for now I'll leave it up as a shameful reminder.