Rewrite Problem3 solution to compute factors first
The solution for Problem3 was unnecessarily expensive because I was computing a list of all prime numbers and then checking if they were factors of the input; prime testing is a more expensive check than checking if x is a factor, and what's more my scheme for computing a list of Prime numbers was itself unnecessarily expensive and complex.

Rewrite the solution to compute a list of potential factors of the input N, and then check if those factors are prime; I also switched to using the builtin big.ProbablyPrime() function for primality testing, as it is both efficient and effictive for any input my program is designed to handle. The new solution completes almost instantly, as opposed to taking minutes previously.