Add Problem5 Solution
Add a Go solution for Problem5. Compute the Least Common Multiple of a set of input numbers by first computing the prime-factorization of each input term, and recording the largest occurrence of each prime factor in a map, then multiplying together the resulting list of prime factors. The solution uses a direct search factorization algorithm so it likely won't scale well to very large input factors, but it does quite well for smaller inputs.